+++ shiki_organizer/cli/task.py
@@ -8,7 +8,7 @@ import click
 from shiki_organizer.actions import get_status
 from shiki_organizer.model import Interval, Task
 import shiki_organizer.actions as actions
-from shiki_organizer.cli.formatting import task_to_str
+from shiki_organizer.cli.formatter import task_to_str
 from shiki_organizer.datetime import period_to_datetime