+++ shiki_organizer/gui/__main__.py
@@ -3,6 +3,7 @@ import sys
 from PySide6.QtWidgets import QApplication, QMainWindow
 from shiki_organizer.gui.ui.main_window import Ui_MainWindow
 from shiki_organizer.gui.tasks_tab import TasksTab
+from shiki_organizer.gui.interval_tab import IntervalTab
 
 
 class MainWindow(QMainWindow, Ui_MainWindow):
@@ -11,6 +12,7 @@ class MainWindow(QMainWindow, Ui_MainWindow):
         self.setupUi(self)
         self.tab_widget.addTab(TasksTab(True), "Today")
         self.tab_widget.addTab(TasksTab(False), "Tasks")
+        self.tab_widget.addTab(IntervalTab(), "Intervals")
 
 
 def main():